Split drivers test files
